cutmix + tests + scripts update (#651)
* cutmix + tests + scripts update

* sh update

* ports update

+++ bin/tests/check_dl.sh
@@ -1,18 +1,10 @@
 #!/usr/bin/env bash
-pip install tifffile #TODO: check if really required
+pip install tifffile  #TODO: check if really required
 
-# gdrive
-#function gdrive_download () {
-#  CONFIRM=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate "https://docs.google.com/uc?export=download&id=$1" -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')
-#  wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$CONFIRM&id=$1" -O $2
-#  rm -rf /tmp/cookies.txt
-#}
-#mkdir -p data
-#gdrive_download 1N82zh0kzmnzqRvUyMgVOGsCoS1kHf3RP ./data/isbi.tar.gz
-#tar -xf ./data/isbi.tar.gz -C ./data/
-
-# aws
 mkdir -p data
+# gdrive
+# gdrive_download 1N82zh0kzmnzqRvUyMgVOGsCoS1kHf3RP ./data/isbi.tar.gz
+# aws
 wget https://catalyst-ai.s3-eu-west-1.amazonaws.com/isbi.tar.gz -O ./data/isbi.tar.gz
 tar -xf ./data/isbi.tar.gz -C ./data/
 
@@ -25,7 +17,12 @@ set -e
 #(set -e; for f in examples/_tests_scripts/z_*.py; do PYTHONPATH=./catalyst:${PYTHONPATH} python "$f"; done)
 
 
-# pipeline 1
+################################  pipeline 00  ################################
+rm -rf ./examples/logs
+
+
+################################  pipeline 01  ################################
+echo 'pipeline 01'
 EXPDIR=./examples/_tests_mnist_stages
 LOGDIR=./examples/logs/_tests_mnist_stages1
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -42,11 +39,12 @@ if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
     exit 1
 fi
 
+cat $LOGFILE
 python -c """
 from safitty import Safict
 metrics = Safict.load('$LOGFILE')
 assert metrics.get('stage1.3', 'loss') < metrics.get('stage1.1', 'loss')
-assert metrics.get('stage1.3', 'loss') < 2.0
+assert metrics.get('stage1.3', 'loss') < 2.1
 """
 
 PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
@@ -56,7 +54,8 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
 rm -rf $LOGDIR
 
 
-# pipeline 2
+################################  pipeline 02  ################################
+echo 'pipeline 02'
 EXPDIR=./examples/_tests_mnist_stages
 LOGDIR=./examples/logs/_tests_mnist_stages1
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -73,11 +72,12 @@ if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
     exit 1
 fi
 
+cat $LOGFILE
 python -c """
 from safitty import Safict
 metrics = Safict.load('$LOGFILE')
 assert metrics.get('stage1.3', 'loss') < metrics.get('stage1.1', 'loss')
-assert metrics.get('stage1.3', 'loss') < 2.0
+assert metrics.get('stage1.3', 'loss') < 2.1
 """
 
 
@@ -89,6 +89,7 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
   --out_dir=${LOGDIR}/:str \
   --out_prefix="/predictions/":str
 
+cat $LOGFILE
 python -c """
 import numpy as np
 data = np.load('${LOGDIR}/predictions/infer.logits.npy')
@@ -98,7 +99,8 @@ assert data.shape == (10000, 10)
 rm -rf $LOGDIR
 
 
-# pipeline 3
+################################  pipeline 03  ################################
+echo 'pipeline 03'
 EXPDIR=./examples/_tests_mnist_stages
 LOGDIR=./examples/logs/_tests_mnist_stages1
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -110,10 +112,24 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
   --logdir=${LOGDIR} \
   --check
 
+if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
+    echo "File $LOGFILE does not exist"
+    exit 1
+fi
+
+cat $LOGFILE
+python -c """
+from safitty import Safict
+metrics = Safict.load('$LOGFILE')
+assert metrics.get('stage1.3', 'loss') < metrics.get('stage1.1', 'loss')
+assert metrics.get('stage1.3', 'loss') < 2.1
+"""
+
 rm -rf ${LOGDIR}
 
 
-# pipeline 4
+################################  pipeline 04  ################################
+echo 'pipeline 04'
 EXPDIR=./examples/_tests_mnist_stages
 LOGDIR=./examples/logs/_tests_mnist_stages1
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -125,10 +141,25 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
   --logdir=${LOGDIR} \
   --check
 
+if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
+    echo "File $LOGFILE does not exist"
+    exit 1
+fi
+
+cat $LOGFILE
+python -c """
+from safitty import Safict
+metrics = Safict.load('$LOGFILE')
+assert metrics.get('stage2.3', 'loss') < metrics.get('stage2.1', 'loss')
+assert metrics.get('stage2.3', 'loss') < 2.1
+"""
+
 rm -rf ${LOGDIR}
 
 
-# pipeline 5
+################################  pipeline 05  ################################
+echo 'pipeline 05'
+# LrFinder
 EXPDIR=./examples/_tests_mnist_stages
 LOGDIR=./examples/logs/_tests_mnist_stages1
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -140,10 +171,24 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
   --logdir=${LOGDIR} \
   --check
 
+if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
+    echo "File $LOGFILE does not exist"
+    exit 1
+fi
+
+cat $LOGFILE
+python -c """
+from safitty import Safict
+metrics = Safict.load('$LOGFILE')
+assert metrics.get('stage2.3', 'loss') < metrics.get('stage2.1', 'loss')
+assert metrics.get('stage2.3', 'loss') < 14.5
+"""
+
 rm -rf ${LOGDIR}
 
 
-# pipeline 6
+################################  pipeline 06  ################################
+echo 'pipeline 06'
 EXPDIR=./examples/_tests_mnist_stages
 LOGDIR=./examples/logs/_tests_mnist_stages_finder
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -160,7 +205,8 @@ kill %1
 rm -rf ${LOGDIR}
 
 
-# pipeline 7
+################################  pipeline 07  ################################
+echo 'pipeline 07'
 EXPDIR=./examples/_tests_mnist_stages2
 LOGDIR=./examples/logs/_tests_mnist_stages2
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -172,10 +218,24 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
   --logdir=${LOGDIR} \
   --check
 
+if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
+    echo "File $LOGFILE does not exist"
+    exit 1
+fi
+
+cat $LOGFILE
+python -c """
+from safitty import Safict
+metrics = Safict.load('$LOGFILE')
+# assert metrics.get('stage1.3', 'loss') < metrics.get('stage1.1', 'loss')
+assert metrics.get('best', 'loss') < 2.35
+"""
+
 rm -rf ${LOGDIR}
 
 
-# pipeline 8
+################################  pipeline 08  ################################
+echo 'pipeline 08'
 EXPDIR=./examples/_tests_mnist_stages2
 LOGDIR=./examples/logs/_tests_mnist_stages2
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -187,10 +247,24 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
   --logdir=${LOGDIR} \
   --check
 
+if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
+    echo "File $LOGFILE does not exist"
+    exit 1
+fi
+
+cat $LOGFILE
+python -c """
+from safitty import Safict
+metrics = Safict.load('$LOGFILE')
+assert metrics.get('stage1.3', 'loss') < metrics.get('stage1.1', 'loss')
+assert metrics.get('stage1.3', 'loss') < 2.35
+"""
+
 rm -rf ${LOGDIR}
 
 
-# pipeline 9
+################################  pipeline 09  ################################
+echo 'pipeline 09'
 EXPDIR=./examples/_tests_mnist_stages2
 LOGDIR=./examples/logs/_tests_mnist_stages2
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -202,10 +276,24 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
   --logdir=${LOGDIR} \
   --check
 
+if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
+    echo "File $LOGFILE does not exist"
+    exit 1
+fi
+
+cat $LOGFILE
+python -c """
+from safitty import Safict
+metrics = Safict.load('$LOGFILE')
+# assert metrics.get('stage1.3', 'loss') < metrics.get('stage1.1', 'loss')
+assert metrics.get('stage1.3', 'loss') < 2.33
+"""
+
 rm -rf ${LOGDIR}
 
 
-# pipeline 10
+################################  pipeline 10  ################################
+echo 'pipeline 10'
 EXPDIR=./examples/_tests_mnist_stages2
 LOGDIR=./examples/logs/_tests_mnist_stages_finder
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -221,17 +309,19 @@ kill %1
 
 rm -rf ${LOGDIR}
 
+################################  pipeline 11  ################################
 # SEGMENTATION
+echo 'pipeline 11 - SEGMENTATION'
 EXPDIR=./examples/_test_segmentation
 LOGDIR=./examples/logs/_test_segmentation
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
 
 ## load the data
-mkdir -p ./examples/_test_segmentation/data
-cd ./examples/_test_segmentation/data/
-download-gdrive 1iYaNijLmzsrMlAdMoUEhhJuo-5bkeAuj segmentation_data.zip
-extract-archive segmentation_data.zip
-cd ../../..
+# mkdir -p ./examples/_test_segmentation/data
+# cd ./examples/_test_segmentation/data/
+# download-gdrive 1iYaNijLmzsrMlAdMoUEhhJuo-5bkeAuj segmentation_data.zip
+# extract-archive segmentation_data.zip
+# cd ../../..
 
 ## train
 PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
@@ -239,6 +329,8 @@ PYTHONPATH=./examples:./catalyst:${PYTHONPATH} \
   --expdir=${EXPDIR} \
   --configs ${EXPDIR}/config.yml ${EXPDIR}/transforms.yml \
   --logdir=${LOGDIR} \
+  --stages/data_params/image_path=./examples/_test_segmentation/data/segmentation_data/train:str \
+  --stages/data_params/mask_path=./examples/_test_segmentation/data/segmentation_data/train_masks:str \
   --check
 
 ## check metrics
@@ -247,6 +339,7 @@ if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
     exit 1
 fi
 
+cat $LOGFILE
 python -c """
 from safitty import Safict
 metrics = Safict.load('$LOGFILE')
@@ -264,7 +357,9 @@ assert loss < 0.2, f'loss must be < 0.2, got {loss}'
 ## remove logs
 rm -rf ./examples/logs/_test_segmentation
 
+################################  pipeline 12  ################################
 # GAN
+echo 'pipeline 12 -  GAN'
 EXPDIR=./examples/mnist_gan
 LOGDIR=./examples/logs/mnist_gan
 LOGFILE=${LOGDIR}/checkpoints/_metrics.json
@@ -281,6 +376,7 @@ if [[ ! (-f "$LOGFILE" && -r "$LOGFILE") ]]; then
     exit 1
 fi
 
+cat $LOGFILE
 python -c """
 from safitty import Safict
 metrics=Safict.load('$LOGFILE')
@@ -307,4 +403,5 @@ assert loss_d < 0.9
 
 rm -rf ${LOGDIR}
 
+################################  pipeline 99  ################################
 rm -rf ./examples/logs
